automatically backed up by learn

+++ lib/bachelor.rb
@@ -54,7 +54,7 @@ def get_average_age_for_season(data, season)
   data.each do |which,contestants|
     if which == season
       contestants.each do |contestant|
-        running_total = running_total + contestant[:age].to_f
+        running_total = running_total + contestant[:age].to_i
       end
       average_age = (running_total/contestants.length)
     end
